Better Bliss phone demo
A long, long time ago, there was this format named gif. While the Internet
still, even to this day, hasn't decided on what pronounciation the word
must be, whether it's pronounced 'j-if' or 'g-if', it was a great format to
show photos on websites.

But gif was abused. Instead of displaying photos, it was forced and
enslaved to forever play looping videos of cute kittens. Gif was tired,
and hoarding all the video frames to display continuous playback was a
heavy load. But people, in general, did not give any care, they simply wanted
to see cute kittens jumping around their monitor while having no regards
for the poor gif standard.

/s

I'm kidding. Let's all use WebM, people!
Conflicts:
	index.html